add grok pattern validation and unique index
